Windows CI: Fix test-unit for daemon
